msm: camera: use devicetree for getting clock info in CPP
CPP needs to use devicetree node to read the clock information.
This is required because, different targets may use different
set of clocks and this change is needed to  help maintain a
generic CPP driver. Fix minor compilation warning in vpe driver.

+++ drivers/media/platform/msm/camera_v2/pproc/vpe/msm_vpe.c
@@ -531,11 +531,11 @@ static int vpe_init_hardware(struct vpe_device *vpe_dev)
 				"vpe", vpe_dev);
 		if (rc < 0) {
 			pr_err("irq request fail! start=%u\n",
-				vpe_dev->irq->start);
+				(uint32_t) vpe_dev->irq->start);
 			rc = -EBUSY;
 			goto unmap_base;
 		} else {
-			VPE_DBG("Got irq! %d\n", vpe_dev->irq->start);
+			VPE_DBG("Got irq! %d\n", (int)vpe_dev->irq->start);
 		}
 	} else {
 		VPE_DBG("Skip requesting the irq since device is booting\n");
